// src/api/models/service-request-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessModel } from './business-model';
// May contain unused imports in some cases
// @ts-ignore
import { ContactInfoModel } from './contact-info-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestClientIdentificationModel } from './service-request-client-identification-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestClientModel } from './service-request-client-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestInteractionModel } from './service-request-interaction-model';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestStatus } from './service-request-status';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestTypeModel } from './service-request-type-model';
// May contain unused imports in some cases
// @ts-ignore
import { UserModel } from './user-model';

/**
 * 
 * @export
 * @interface ServiceRequestModel
 */
export interface ServiceRequestModel {
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestModel
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ServiceRequestModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {ServiceRequestStatus}
     * @memberof ServiceRequestModel
     */
    'status'?: ServiceRequestStatus;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestModel
     */
    'typeId'?: number;
    /**
     * 
     * @type {ServiceRequestTypeModel}
     * @memberof ServiceRequestModel
     */
    'type'?: ServiceRequestTypeModel;
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestModel
     */
    'businessOwnerId'?: number | null;
    /**
     * 
     * @type {BusinessModel}
     * @memberof ServiceRequestModel
     */
    'businessOwner'?: BusinessModel;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestModel
     */
    'responsibleUserId'?: string | null;
    /**
     * 
     * @type {UserModel}
     * @memberof ServiceRequestModel
     */
    'responsibleUser'?: UserModel;
    /**
     * 
     * @type {ServiceRequestClientModel}
     * @memberof ServiceRequestModel
     */
    'client'?: ServiceRequestClientModel;
    /**
     * 
     * @type {ServiceRequestClientIdentificationModel}
     * @memberof ServiceRequestModel
     */
    'clientIdentification'?: ServiceRequestClientIdentificationModel;
    /**
     * 
     * @type {Array<ContactInfoModel>}
     * @memberof ServiceRequestModel
     */
    'clientContactInfo'?: Array<ContactInfoModel> | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestModel
     */
    'formType'?: string | null;
    /**
     * 
     * @type {any}
     * @memberof ServiceRequestModel
     */
    'formData'?: any | null;
    /**
     * 
     * @type {Array<ServiceRequestInteractionModel>}
     * @memberof ServiceRequestModel
     */
    'interactions'?: Array<ServiceRequestInteractionModel> | null;
}